pyproject.toml
[project]
name = "provider-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ruff>=0.15.6",
    "httpx>=0.28",
    "beautifulsoup4>=4.14.3",
    "pytest-playwright>=0.7",
    "pytest>=9.0",
    "python-dotenv>=1.2.2",
    "pre-commit>=4.5.1",
    "boto3>=1.42.71",
]

[tool.setuptools]
packages = ["provider_data"]

[tool.uv]
package = true
